# stage1_gateway_openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ForgeAI - Api Gateway",
    "description": "API Gateway single public entry point providing authentication, resilience, and metrics.\n\n*Part of the ForgeAI Local AI Software Engineering Platform.*",
    "version": "0.1.0",
    "x-service-name": "api-gateway"
  },
  "paths": {
...
    "/api/v1/repositories/browse": {
      "get": {
        "tags": [
          "Repositories"
        ],
        "summary": "Browse Local Filesystem Directories",
        "description": "Returns available drives, current path, parent, and subdirectories for folder navigation.",
        "operationId": "browse_filesystem_api_v1_repositories_browse_get",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Filesystem directory listing retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse_FilesystemBrowseResponse_"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/repositories": {
      "get": {
        "tags": [
          "Repositories"
        ],
        "summary": "List Registered Repositories",
        "description": "Returns all registered repository metadata records for the authenticated user.",
        "operationId": "list_repositories_api_v1_repositories_get",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Repositories list retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse_list_dict__"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/repositories/open": {
      "post": {
        "tags": [
          "Repositories"
        ],
        "summary": "Open Local Repository",
        "description": "Registers and inspects a local repository directory.",
        "operationId": "open_repository_api_v1_repositories_open_post",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OpenRepoRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Repository opened",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse_RepoStatusResponse_"
                }
              }
            }
          },
          "400": {
            "description": "Invalid directory path",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/repositories/clone": {
      "post": {
        "tags": [
          "Repositories"
        ],
        "summary": "Clone Git Repository",
        "description": "Clones a remote repository URL locally.",
        "operationId": "clone_repository_api_v1_repositories_clone_post",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CloneRepoRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Cloning started",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse_RepoStatusResponse_"
                }
              }
            }
          },
          "400": {
            "description": "Invalid git URL",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/repositories/{id}/index": {
      "post": {
        "tags": [
          "Repositories"
        ],
        "summary": "Index Repository Vector RAG",
        "description": "Parses AST symbols and generates Qdrant embeddings for code search.",
        "operationId": "index_repository_api_v1_repositories__id__index_post",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IndexRepoRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Indexing started",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse_RepoStatusResponse_"
                }
              }
            }
          },
          "404": {
            "description": "Repository not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/repositories/{id}/tree": {
      "get": {
        "tags": [
          "Repositories"
        ],
        "summary": "Get Repository File Tree",
        "description": "Returns the filtered source code directory file tree hierarchy.",
        "operationId": "get_repository_tree_api_v1_repositories__id__tree_get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Tree hierarchy retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse_list_RepoTreeNode__"
                }
              }
            }
          },
          "404": {
            "description": "Repository not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/repositories/{id}/status": {
      "get": {
        "tags": [
          "Repositories"
        ],
        "summary": "Get Repository Indexing Status",
        "description": "Retrieves current parsing and indexing progress.",
        "operationId": "get_repository_status_api_v1_repositories__id__status_get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Status retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse_RepoStatusResponse_"
                }
              }
            }
          },
          "404": {
            "description": "Repository not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/repositories/{id}/file": {
      "get": {
        "tags": [
          "Repositories"
        ],
        "summary": "Get File Content",
        "description": "Reads content of a file within the repository by its relative path.",
        "operationId": "get_repository_file_api_v1_repositories__id__file_get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          },
          {
            "name": "path",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Path"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/repositories/{id}/files": {
      "post": {
        "tags": [
          "Repositories"
        ],
        "summary": "Create or Update File",
        "description": "Writes text content to a target file.",
        "operationId": "create_or_update_file_api_v1_repositories__id__files_post",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FileWriteRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CloneRepoRequest": {
        "properties": {
          "url": {
            "type": "string",
            "title": "Url",
            "description": "Git repository remote URL",
            "example": "https://github.com/org/repo.git"
          },
          "destination_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Destination Path",
            "description": "Optional target directory"
          }
        },
        "type": "object",
        "required": [
          "url"
        ],
        "title": "CloneRepoRequest",
        "description": "Request payload to clone a remote git repository."
      },
...
      "ErrorResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "description": "False for failed requests",
            "default": false
          },
          "request_id": {
            "type": "string",
            "title": "Request Id",
            "description": "Correlation ID for request tracing",
            "default": "-"
          },
          "trace_id": {
            "type": "string",
            "title": "Trace Id",
            "description": "Distributed trace ID",
            "default": "-"
          },
          "timestamp": {
            "type": "string",
            "title": "Timestamp",
            "description": "ISO 8601 UTC timestamp"
          },
          "code": {
            "type": "string",
            "title": "Code",
            "description": "Machine-readable error code",
            "default": "INTERNAL_ERROR"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Human-readable error description"
          },
          "service": {
            "type": "string",
            "title": "Service",
            "description": "Originating service name",
            "default": "api-gateway"
          },
          "error": {
            "$ref": "#/components/schemas/ErrorDetails",
            "description": "Structured error details"
          }
        },
        "type": "object",
        "required": [
          "message",
          "error"
        ],
        "title": "ErrorResponse",
        "description": "Unified error response wrapper conforming to production contract."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OpenRepoRequest": {
        "properties": {
          "path": {
            "type": "string",
            "title": "Path",
            "description": "Absolute local directory path",
            "example": "E:/Repos/my-app"
          }
        },
        "type": "object",
        "required": [
          "path"
        ],
        "title": "OpenRepoRequest",
        "description": "Request payload to open a local repository directory."
      },
...
      "SuccessResponse_FilesystemBrowseResponse_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "description": "True for successful requests",
            "default": true
          },
          "request_id": {
            "type": "string",
            "title": "Request Id",
            "description": "Correlation ID for request tracing",
            "default": "-"
          },
          "trace_id": {
            "type": "string",
            "title": "Trace Id",
            "description": "Distributed trace ID",
            "default": "-"
          },
          "timestamp": {
            "type": "string",
            "title": "Timestamp",
            "description": "ISO 8601 UTC timestamp"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Response message",
            "default": "Operation completed successfully"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FilesystemBrowseResponse"
              },
              {
                "type": "null"
              }
            ],
            "description": "Payload data"
          }
        },
        "type": "object",
        "title": "SuccessResponse[FilesystemBrowseResponse]"
      },
...
      "SuccessResponse_RepoStatusResponse_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "description": "True for successful requests",
            "default": true
          },
          "request_id": {
            "type": "string",
            "title": "Request Id",
            "description": "Correlation ID for request tracing",
            "default": "-"
          },
          "trace_id": {
            "type": "string",
            "title": "Trace Id",
            "description": "Distributed trace ID",
            "default": "-"
          },
          "timestamp": {
            "type": "string",
            "title": "Timestamp",
            "description": "ISO 8601 UTC timestamp"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Response message",
            "default": "Operation completed successfully"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RepoStatusResponse"
              },
              {
                "type": "null"
              }
            ],
            "description": "Payload data"
          }
        },
        "type": "object",
        "title": "SuccessResponse[RepoStatusResponse]"
      },
...
      "SuccessResponse_list_dict__": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "description": "True for successful requests",
            "default": true
          },
          "request_id": {
            "type": "string",
            "title": "Request Id",
            "description": "Correlation ID for request tracing",
            "default": "-"
          },
          "trace_id": {
            "type": "string",
            "title": "Trace Id",
            "description": "Distributed trace ID",
            "default": "-"
          },
          "timestamp": {
            "type": "string",
            "title": "Timestamp",
            "description": "ISO 8601 UTC timestamp"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Response message",
            "default": "Operation completed successfully"
          },
          "data": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data",
            "description": "Payload data"
          }
        },
        "type": "object",
        "title": "SuccessResponse[list[dict]]"
      },
...
    }
  }
}